+++ R/chkpt_brms.R
@@ -10,8 +10,15 @@
 #' @param data An object of class \code{data.frame} (or one that can be coerced
 #'   to that class) containing data of all variables used in the model.
 #'
+#' @param iter_adaptation (positive integer) The number of iterations in the
+#'   initial warmup, which are used for the adaptation of the step size and 
+#'   inverse mass matrix. This is equivalent to the traditional warmup stage.
+#'   Checkpointing will begin only after this stage is complete.
+#'
 #' @param iter_warmup (positive integer) The number of warmup iterations to run
-#'   per chain (defaults to 1000).
+#'   per chain after the adaptation stage (defaults to 1000). During this stage
+#'   the step size and inverse mass matrix are fixed to the values found during
+#'   the adaptation stage. There is no further adaptation performed.
 #'
 #' @param iter_sampling (positive integer) The number of post-warmup iterations
 #'   to run per chain (defaults to 1000).
@@ -20,12 +27,6 @@
 #'   checkpoint. Note that \code{iter_sampling} is divided by
 #'   \code{iter_per_chkpt} to determine the number of checkpoints. This must
 #'   result in an integer (if not, there will be an error).
-#'
-#' @param iter_typical (positive integer) The number of iterations in the
-#'   initial warmup, which finds the so-called typical set. This is an initial
-#'   phase, and not included in \code{iter_warmup}. Note that a large enough
-#'   value is required to ensure convergence (defaults to 150).
-#'
 #'
 #' @param parallel_chains (positive integer) The \emph{maximum number} of MCMC
 #'   chains to run in parallel. If parallel_chains is not specified then the
@@ -44,9 +45,6 @@
 #' @param control A named list of parameters to control the sampler's behavior.
 #'   It defaults to NULL so all the default values are used. For a comprehensive
 #'   overview see \code{\link[rstan]{stan}}.
-#'
-#' @param brmsfit Logical. Should a \code{brmsfit} object be returned (defaults
-#'   to \code{TRUE}).
 #'
 #' @param seed (positive integer). The seed for random number generation to make
 #'   results reproducible.
@@ -73,16 +71,14 @@
 #'   \code{\link[brms]{brmsfamily}} (e.g., \code{family = poisson())}, data2,
 #'   custom_families, etc.
 #'
-#' @return An object of class \code{brmsfit} (with \code{brmsfit = TRUE}) or
-#'   \code{chkpt_brms} (with \code{brmsfit = FALSE}).
-#'
+#' @return An object of class \code{brmsfit}
 #' @note
 #'
 #' A folder specified by \code{path} is created with four subfolders:
 #'
 #' \itemize{
 #'
-#' \item \strong{cmd_fit}: The cmdstanr fittted models (one for each checkpoint).
+#' \item \strong{cmd_output}: The cmdstanr output_files (one for each checkpoint and chain).
 #'
 #' \item \strong{cp_info}: Mass matrix, step size, and initial values for
 #'                       next checkpoint (last iteration from previous checkpoint).
@@ -167,15 +163,14 @@
 #' }
 chkpt_brms <- function(formula,
                        data,
+                       iter_adaptation = 150,
                        iter_warmup = 1000,
                        iter_sampling = 1000,
                        iter_per_chkpt = 100,
-                       iter_typical = 150,
                        parallel_chains = 4,
                        threads_per = 1,
                        chkpt_progress = TRUE,
                        control = NULL,
-                       brmsfit = TRUE,
                        seed = 1,
                        stop_after = NULL,
                        reset = FALSE,
@@ -184,18 +179,18 @@ chkpt_brms <- function(formula,
   # TODO: MASSIVElY SIMPLIFY AND REFACTOR ALL CODE BELOW AFTER THE HOTFIX IS OUT
   stan_phase = ""
   args <- c(as.list(environment()), list(...))
+  if (!is.null(args$iter_typical)) {
+    warning("iter_typical is deprecated. Please use iter_adaptation instead.")
+    iter_adaptation <- args$iter_typical
+  }
+  
 
   withr::defer({
     if (stan_phase %in% c("sample", "complete")) {
       if (is.null(returnValue())) {
-        message("\nSampling aborted. You can examine the results or continue sampling by rerunning the same code.")
-        return(return_object(
-          brmsfit = brmsfit,
-          formula = formula,
-          data = data,
-          path = path,
-          ...
-        ))
+        message("\nSampling aborted. You can examine the results or continue", 
+                " sampling by rerunning the same code.")
+        return(make_brmsfit(formula = formula, data = data, path = path, ...))
       }
     } else {
       message("\nInterupted before or during warmup. No samples available.")
@@ -221,10 +216,8 @@ chkpt_brms <- function(formula,
 
   # TODO: DON"T REMAKE STAN CODE AND DATA IF EXECUTABLE EXISTS
   if (threads_per == 1) {
-    stan_data <-
-      brms::make_standata(formula = formula, data = data, ...)
-    stan_code <-
-      brms::make_stancode(formula = formula, data = data, ...)
+    stan_data <- brms::make_standata(formula = formula, data = data, ...)
+    stan_code <- brms::make_stancode(formula = formula, data = data, ...)
   } else {
     stan_data <- brms::make_standata(
       formula = formula,
@@ -294,7 +287,7 @@ chkpt_brms <- function(formula,
       seed = seed,
       control = control,
       model = stan_m3$sample,
-      iter_typical = iter_typical,
+      iter_adaptation = iter_adaptation,
       cmd_args = cmd_args,
       progress = chkpt_progress
     )
@@ -315,14 +308,9 @@ chkpt_brms <- function(formula,
   }
 
   if (last_chkpt == total_chkpts) {
+    stan_phase <- "complete"
     message("Checkpointing complete")
-    return(return_object(
-      brmsfit = brmsfit,
-      formula = formula,
-      data = data,
-      path = path,
-      ...
-    ))
+    return(make_brmsfit(formula = formula, data = data, path = path, ...))
   } else {
     # TODO: remove unnecessary else clause
     cp_seq <- seq(last_chkpt + 1, total_chkpts)
@@ -332,61 +320,33 @@ chkpt_brms <- function(formula,
     if (!is.null(stop_after) && i > stop_at_checkpoint) {
       message("Stopping after ", stop_at_checkpoint, " checkpoints")
       if (i > warmup_chkpts + 1) {
-        return(return_object(
-          brmsfit = brmsfit,
-          formula = formula,
-          data = data,
-          path = path,
-          ...
-        ))
+        return(make_brmsfit(formula = formula, data = data, path = path, ...))
       }
       stan_phase <- "warmup"
       return(invisible(NULL))
     }
 
-    if (i <= warmup_chkpts) {
-      stan_phase <- "warmup"
-
-      sample_chunk <- chkpt_sample(
-        model = stan_m3$sample,
-        cmd_args = cmd_args,
-        control = control,
-        progress = chkpt_progress,
-        cp_cmd_args = cp_cmd_args(
-          seed = seed + i,
-          phase = stan_phase,
-          stan_state = stan_state,
-          iter_per_chkpt = iter_per_chkpt
-        )
+    stan_phase <- c("warmup", "sample")[1 + (i > warmup_chkpts)]
+    sample_chunk <- chkpt_sample(
+      model = stan_m3$sample,
+      cmd_args = cmd_args,
+      control = control,
+      progress = chkpt_progress,
+      cp_cmd_args = cp_cmd_args(
+        seed = seed + i,
+        phase = stan_phase,
+        stan_state = stan_state,
+        iter_per_chkpt = iter_per_chkpt,
+        path = path,
+        checkpoint = i
       )
-    } else {
-      stan_phase <- "sample"
-
-      sample_chunk <- chkpt_sample(
-        model = stan_m3$sample,
-        cmd_args = cmd_args,
-        control = control,
-        progress = chkpt_progress,
-        cp_cmd_args = cp_cmd_args(
-          seed = seed + i,
-          phase = stan_phase,
-          stan_state = stan_state,
-          iter_per_chkpt = iter_per_chkpt
-        )
-      )
-
+    )
+    
+    if (stan_phase == "sample") {
       sample_chunk$save_object(paste0(
         path, "/cp_samples/",
         "samples_", i, ".rds"
       ))
-
-      # TODO: THIS CAN BE OPTIMIZED BY DOING ONLY ONCE AT THE END
-      if (brmsfit) {
-        saveRDS(
-          object = rstan::read_stan_csv(sample_chunk$output_files()),
-          file = paste0(path, "/cmd_fit/cmd_fit_", i, ".rds")
-        )
-      }
     }
 
     stan_state <- extract_stan_state(sample_chunk, stan_phase)
@@ -401,30 +361,9 @@ chkpt_brms <- function(formula,
     if (i == total_chkpts) {
       message("Checkpointing complete")
       stan_phase <- "complete"
-      return(return_object(
-        brmsfit = brmsfit,
-        formula = formula,
-        data = data,
-        path = path,
-        ...
-      ))
+      return(make_brmsfit(formula = formula, data = data, path = path, ...))
     }
   }
-}
-
-
-return_object <- function(brmsfit, ...) {
-  dots <- list(...)
-  args <- dots$args
-  dots$args <- NULL
-  if (brmsfit) {
-    out <- brms::do_call(make_brmsfit, dots)
-  } else {
-    out <- list(args = args)
-    class(out) <- "chkpt_brms"
-  }
-  out$path <- dots$path
-  out
 }